Changed logic to retrieve tooltip for a facet by turning the facet into a FacetType enum (previously known as FacetsToTooltipMapping). Fixed indentation.
